refactor(web): space-create → moim-create 네이밍 통일
