+++ apps/web/src/_pages/moim-create/assets/ico-project.png
