+++ apps/web/src/_pages/moim-create/assets/ico-title.png
