+++ apps/web/src/_pages/moim-create/assets/index.ts
