+++ apps/web/src/features/moim-create/ui/theme-color-select.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { SelectBox } from "@ui/components/ui/selectbox";
-import { THEME_COLORS } from "@/features/space-create/constants";
+import { THEME_COLORS } from "@/features/moim-create/constants";
 
 interface ThemeColorSelectProps {
   value?: string;
